+++ app/src/test/java/io/appium/uiautomator2/model/settings/WaitForSelectorTimeoutTests.java
@@ -34,7 +34,6 @@ import static org.mockito.Mockito.when;
 @RunWith(PowerMockRunner.class)
 @PrepareForTest({Configurator.class})
 public class WaitForSelectorTimeoutTests {
-
     private WaitForSelectorTimeout waitForSelectorTimeout;
 
     @Mock
@@ -60,13 +59,13 @@ public class WaitForSelectorTimeoutTests {
 
     @Test
     public void shouldBeAbleToSetSelectorTimeout() {
-        waitForSelectorTimeout.update(123);
-        verify(configurator).setWaitForSelectorTimeout(123);
+        waitForSelectorTimeout.update(123L);
+        verify(configurator).setWaitForSelectorTimeout(123L);
     }
 
     @Test
     public void shouldBeAbleToGetSelectorTimeout() {
-        when(configurator.getWaitForSelectorTimeout()).thenReturn((long) 123);
+        when(configurator.getWaitForSelectorTimeout()).thenReturn(123L);
         Assert.assertEquals(Long.valueOf(123), waitForSelectorTimeout.getValue());
     }
 }
